Revert "use callback for apk handling,  and fix uri processing"
This reverts commit 7edbe74af1a9e732687e32641bf77f7e8fe521e3.
