+++ app/src/main/java/com/fieldbook/tracker/utilities/VersionChecker.java
@@ -4,6 +4,8 @@ import android.app.AlertDialog;
 import android.app.ProgressDialog;
 import android.content.Context;
 import android.content.DialogInterface;
+import android.content.Intent;
+import android.net.Uri;
 import android.os.AsyncTask;
 import android.util.Log;
 
@@ -16,7 +18,6 @@ import org.json.JSONObject;
 import org.phenoapps.utils.BaseDocumentTreeUtil;
 
 import java.io.BufferedReader;
-import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
@@ -27,6 +28,7 @@ import java.net.URL;
 
 import javax.annotation.Nullable;
 
+
 public class VersionChecker extends AsyncTask<Void, Void, String> {
     private static final String GITHUB_API_URL = "https://api.github.com/repos/{owner}/{repo}/releases/latest";
     private static final String TAG = "VersionChecker";
@@ -34,10 +36,6 @@ public class VersionChecker extends AsyncTask<Void, Void, String> {
     private final String owner;
     private final String repo;
     private final WeakReference<Context> context; // Add context field
-    private VersionCheckerListener listener;
-    public void setListener(VersionCheckerListener listener) {
-        this.listener = listener;
-    }
 
     @Nullable
     protected Context getContext() {
@@ -50,7 +48,6 @@ public class VersionChecker extends AsyncTask<Void, Void, String> {
         this.currentVersion = currentVersion;
         this.owner = owner;
         this.repo = repo;
-        this.listener = listener;
     }
 
     @Override
@@ -77,12 +74,6 @@ public class VersionChecker extends AsyncTask<Void, Void, String> {
             Log.e(TAG, "Error occurred while checking for updates: " + e.getMessage());
         }
         return null;
-    }
-
-    private void onApkDownloaded(File apkFile) {
-        if (listener != null) {
-            listener.onApkDownloaded(apkFile);
-        }
     }
 
     @Override
@@ -222,20 +213,67 @@ public class VersionChecker extends AsyncTask<Void, Void, String> {
         @Override
         protected void onPostExecute(DocumentFile apkFile) {
             super.onPostExecute(apkFile);
+
             Context context = getContext();
+
             if (context != null) {
+
+                // Dismiss the progress dialog
                 progressDialog.dismiss();
+
                 if (apkFile != null) {
-                    Log.d("Notifying AboutActivity", "triggering callback");
-                    // Notify the listener that the APK download is completed
-                    if (listener != null) {
-                        Log.d("using listner", "triggering callback via listener");
-                        listener.onApkDownloaded(new File(apkFile.getUri().getPath()));
-                    }
+                    // // Install the APK
+                    // Uri apkUri = apkFile.getUri();
+                    // Intent installIntent = new Intent(Intent.ACTION_VIEW);
+                    // installIntent.setDataAndType(apkUri, "application/vnd.android.package-archive");
+                    // installIntent.addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION);
+                    // installIntent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+                    // context.startActivity(installIntent);
+
+                    // Display a message instructing the user to install the APK and open the directory
+                    AlertDialog.Builder builder = new AlertDialog.Builder(context, R.style.AlertDialogStyle);
+                    builder.setTitle("Download Completed");
+                    builder.setMessage("The update file has been successfully downloaded to the updates directory.\n\nOpen the directory and double click the APK file to install it.");
+
+                    builder.setPositiveButton("Open Directory", new DialogInterface.OnClickListener() {
+                        @Override
+                        public void onClick(DialogInterface dialog, int which) {
+                            // Open the directory where the APK was saved
+                            openDownloadDirectory(apkFile);
+                            dialog.dismiss();
+                        }
+                    });
+
+                    builder.setNegativeButton(android.R.string.cancel, new DialogInterface.OnClickListener() {
+                        @Override
+                        public void onClick(DialogInterface dialog, int which) {
+                            dialog.dismiss();
+                        }
+                    });
+
+                    builder.show();
                 }
             }
         }
     }
+
+    private void openDownloadDirectory(DocumentFile apkFile) {
+    Context context = getContext();
+    if (context != null) {
+        DocumentFile parentDirectory = apkFile.getParentFile();
+        if (parentDirectory != null) {
+            Uri uri = parentDirectory.getUri();
+            Intent intent = new Intent(Intent.ACTION_GET_CONTENT);
+            intent.setDataAndType(uri, "*/*");
+            intent.addCategory(Intent.CATEGORY_OPENABLE);
+            intent.addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION);
+            intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+            if (intent.resolveActivity(context.getPackageManager()) != null) {
+                context.startActivity(intent);
+            }
+        }
+    }
+}
 
     private boolean isNewerVersion(String currentVersion, String latestVersion) {
         // Split the version strings into their components
