--- app/src/main/java/com/fieldbook/tracker/utilities/VersionCheckerListener.java
@@ -1,7 +1,0 @@
-package com.fieldbook.tracker.utilities;
-
-import java.io.File;
-
-public interface VersionCheckerListener {
-    void onApkDownloaded(File apkFile);
-}
