+++ app/src/main/res/xml/provider_paths.xml
@@ -1,10 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
-<paths xmlns:android="http://schemas.android.com/apk/res/android">
+<paths>
     <external-path
         name="fieldBook"
         path="." />
     <cache-path
         name="cache"
         path="."/>
-    <root-path name="root" path="." />
 </paths>
